issue4-1 Fat Fragment の回避：同一ファイル内に複数のクラスが存在することを回避
